requirements.txt
uptick==0.2.5
bitshares==0.7.0
graphenelib==1.4.0
PyYAML
pypubsub
matplotlib
numpy